fix(lua): fix test_crucible_log and clippy warnings
- Fix crucible.log metatable to preserve function call capability
  - register_log_levels now wraps the original log function in __call
  - Handles metatable __call convention (table as first arg)
  - Allows both crucible.log(...) and crucible.log.levels.INFO

- Fix clippy warnings:
  - Remove unnecessary cast in hooks.rs (raw_len() already returns usize)
  - Handle set_metatable Result properly with ?

- Add test for cru.fmt() functionality (3 new tests)

All tests now pass (411 tests)
All clippy warnings resolved

+++ crates/crucible-lua/src/hooks.rs
@@ -29,7 +29,7 @@ pub fn register_hooks_module(lua: &Lua, crucible: &Table) -> LuaResult<()> {
             .get("on_session_start")
             .unwrap_or_else(|_| lua.create_table().unwrap());
 
-        let len = session_start_hooks.raw_len() as usize;
+        let len = session_start_hooks.raw_len();
         session_start_hooks.raw_set(len + 1, key)?;
 
         hooks_table.set("on_session_start", session_start_hooks)?;
